+++ src/App/VRChatContentManager.App/ViewModels/IAppWindow.cs
@@ -4,4 +4,5 @@ public interface IAppWindow
 {
     void SetPin(bool isPinned);
     bool IsPinned();
+    void Activate();
 }