+++ src/App/VRChatContentManager.App/ViewModels/MainWindowViewModel.cs
@@ -1,12 +1,6 @@
-﻿using System;
-using CommunityToolkit.Mvvm.ComponentModel;
-using Microsoft.Extensions.DependencyInjection;
-using VRChatContentManager.App.Pages.GettingStarted;
+﻿using CommunityToolkit.Mvvm.ComponentModel;
 using VRChatContentManager.App.Services;
 using VRChatContentManager.App.ViewModels.Pages;
-using VRChatContentManager.App.ViewModels.Pages.GettingStarted;
-using VRChatContentManager.Core.Settings;
-using VRChatContentManager.Core.Settings.Models;
 
 namespace VRChatContentManager.App.ViewModels;
 
@@ -15,6 +9,8 @@ public partial class MainWindowViewModel : ViewModelBase, INavigationHost, IAppW
     [ObservableProperty] public partial PageViewModelBase? CurrentPage { get; private set; }
 
     [ObservableProperty] public partial bool Pinned { get; private set; }
+
+    public event EventHandler? RequestActivate;
 
     private readonly NavigationService _navigationService;
     private readonly DialogService _dialogService;
@@ -32,7 +28,7 @@ public partial class MainWindowViewModel : ViewModelBase, INavigationHost, IAppW
         _navigationService.Register(this);
 
         _navigationService.Navigate<BootstrapPageViewModel>();
-        
+
         appWindowService.Register(this);
     }
 
@@ -43,4 +39,9 @@ public partial class MainWindowViewModel : ViewModelBase, INavigationHost, IAppW
 
     public void SetPin(bool isPinned) => Pinned = isPinned;
     public bool IsPinned() => Pinned;
+
+    public void Activate()
+    {
+        RequestActivate?.Invoke(this, EventArgs.Empty);
+    }
 }